Add template C++ classes
The dummy library contains two C++ classes, of which one is public and one
is private. The template classes print messages once they are instantiated
and destroyed.

The installation location of the template-library follows the
"GNUInstallDirs" CMake pattern:
* The shared library binary in deployed in ${LIBDIR}
* The public library headers are deployed in ${INCLUDEDIR}/template-library

A template-library.pc pkgconfig file is created and installed along with
the shared library and public headers. While developing a service
application, a developer can check whether the template-library is
installed on the system by using the pkg-config module for CMake.

Signed-off-by: Gordan Markuš <gordan.markus@pelagicore.com>