CMakeLists.txt: don't use CMAKE_INSTALL_FULL_LIBDIR
Because of a CMake issue [1], CMAKE_INSTALL_FULL_LIBDIR does not point
to a location under CMAKE_STAGING_PREFIX, but defaults to
"/usr/local/lib".
CMAKE_INSTALL_LIBDIR on the other hand is properly taking the value of
CMAKE_STAGING_PREFIX into account, and can be safely used as in practice
it's always an absolute path.

[1]: https://gitlab.kitware.com/cmake/cmake/issues/17340